Fix a duplicate ID for TestTool in the SCMTools fixture.
The SCMTools tests fixture had a duplicate database ID for `TestTool`,
which is used for unit tests. This didn't impact the old fixture loading
code but outright breaks the new code. This change bumps up the ID of
this tool, preventing the conflict, and orders all the other tools to
help prevent this sort of thing from happening in the future.