Enable DVCS by default.
The DVCS feature is now pretty stable, and we want to make sure that 4.0
ships with it turned on so that users can play with it and run it
through its paces. This change turns the DVCS feature flag on.

Testing Done:
Checked locally that the DVCS UI was showing up with this enabled.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10390/