Fix a broken unit test due to a query optimization.
As part of the DVCS work, the review request API resource ended up with
an extra LocalSite-related query, necessitating the bumping of the
number of expected queries in the unit test.

The recent query cleanup removed the need for that extra lookup,
utilizing pre-queried data instead. This change reduces the query count
expected in the test accordingly, bringing it back to the same number we
had in Review Board 3.0.x.