Replace old selectors with smart objects selector widget in the admin UI
Now that `RelatedObjectedWidget` is in Djblets, we import it from there
and extend it in the `RelatedUserWidget`. This change also creates
`RelatedRepositoryWidget` and `RelatedGroupWidget`, along with relevant
HTML and JS files, which extends the `RelatedObjectWidget`.
`RelatedGroupWidget` has the ability to only display groups that are
invite only. These are now used in the places where the admin UI needs
to allow people to select related objects.

Testing Done:
- Ran Python unit tests.
- Ran JS unit tests.
- Tested the default reviewer, repository, and review group forms in the
  admin UI.

Reviewed at https://reviews.reviewboard.org/r/10215/