Clean up classes and styles for the admin sidebar
The admin sidebar was kind of a mess. It used the class names and styles
from the admin widgets, but then did a bunch of overrides and other
element-specific styles. This change breaks it apart and makes it so that
future redesign work will be much easier.

This involves several pieces:
- Renaming the template so it doesn't claim to be a widget.
- Switching away from the `.admin-widget` class and associated sub-element
  classes, and moving towards our new CSS naming standards.
- Defining new rules for the new classes, bringing across only those things
  which are necessary for the look of the sidebar.
- Eliminating style rules specific to each section of the sidebar and making
  those general.

There is one visual change in here, removing the text from the "Add"
links within the "Manage" section. This text really doesn't add anything
to the user experience, and it cluttered up both the visual appearance
and the style rules. I've moved this into a tooltip instead (and changed
it to say "Add new").

Testing Done:
Verified the visual appearance and behavior of the admin sidebar.

Reviewed at https://reviews.reviewboard.org/r/10407/