Rename the admin widget classes.
This change renames the admin widget classes to conform to our new
standard for CSS, and reorganizes a lot of the CSS to match. There's
only one small visual change, switching the background of the "Add Small
Widgets" button to be white to match the appearance of the "Add Large
Widgets" button.

Testing done:
Verified the appearance of the admin dashboard.

Reviewed at https://reviews.reviewboard.org/r/10414/