Fix inserting the help text element for paths in the repository form.
The repository form allows SCMTools to provide help text for the Path
field, and does the work of managing the element, adding or removing it
as needed. However, it was positioning this just outside of the `field`
container. That didn't matter so much currently for the Django or
RBCommons admin UIs, but the position was incorrect and could lead to
some styling problems.

This fixes the positioning so that it's inserted into the `field`
container.

Testing Done:
Checked the field elements in the RB Admin UI and RBC Team Admin UI.
Verified the `.help` element was inserted into `.field`.

Reviewed at https://reviews.reviewboard.org/r/10404/