Fix a crash on review requests without diffs.
The new commit list field made an assumption that all review requests
had diffsets associated with it, and would crash when the diffset came
back as `None`. This is a simple fix that handles the cas of an empty
diffset more gracefully.

Testing Done:
Loaded a review request that didn't have a diffset. Saw that the page
loaded without crashing.

Reviewed at https://reviews.reviewboard.org/r/10412/