Improve documentation and settings names for e-mail.
Our e-mail settings and docs were pretty out-of-date, and a bit
misleading. Some of the settings names had inconsistent capitalization,
and it wasn't always clear who would be receiving the e-mails. The
documentation was a bit worse, leaving out most of the settings and
improperly describing how to avoid suspicious e-mails.

This modernizes all this, renaming some of the settings for clarity and
consistency, and rewriting much of these docs to reflect modern behavior
and configuration for e-mail.

Testing Done:
Built the docs. Checked for build, spelling, and link issues.

Reviewed at https://reviews.reviewboard.org/r/10420/