Add a Follow menu for keeping up with Review Board updates.
This introduces a new "Follow" menu to the top-right of the pages on
Review Board, alongside "Support". It provides a menu that links to our
news feed, Twitter, Facebook, Reddit, and YouTube channel. The idea is
to get users who may use Review Board but not be aware of any updates,
announcements, or discussions around it to have an easy way to find out
about them and follow them.

For those who *really* don't want this enabled (for instance, if it
clashes with an internally-defined menu that might confuse users), this
menu can be disabled through `settings.DISABLE_FOLLOW_MENU`.

This change also introduces a long-overdue CSS modification to make the
menus left-aligned instead of right-aligned. The right-aligned was a
style choice in the rewrite of the header bar UI that really didn't work
for anything more than a couple small items per menu. We already
override this in RBCommons, and with the ability to have custom things
in these menus, left-aligned just makes more sense.

Testing Done:
Tested all the links in the menu.

Reviewed at https://reviews.reviewboard.org/r/10421/