Fix a JavaScript crash in the diff viewer.
Some recent work to convert `RB.PaginationView` to ES6 resulted in a
`const` variable that was defined in one scope and accessed in another.
This resulted in a problem building URLs needed for pagination and
switching commits.

This change fixes up the scoping issues, and also fixes a check that
resulted in the wrong variable being accessed when determining the kind
of result to return.

Testing Done:
Verified the regression, and tested with the fix. URL building and
links worked.

JavaScript unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10424/