Fix a crash with missing review request IDs in the close-submitted hooks.
The common `close_all_review_requests()` function for close-submitted
hooks attempted to log any review request IDs referenced in commits that
couldn't be found in the database. Part of this involved building a map
of the ones that could be found, but the code to do this was broken,
causing a crash.

This change fixes the crash, and adds new unit tests for the affected
services that verify things work correctly when encountering missing
review request IDs.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10434/