Fix issues reloading the Issue Summary Table.
The recent work on the Issue Summary Table caused a regression when
attempting to reload the table after an automated review. The JavaScript
view was setting the element as a child of `#issue-summary`, but when
reloading, we'd replace the view's element with the entire HTML of the
table, _including_ the `#issue-summary` container. This caused a strange
nesting issue and broke element lookups, along with the rest of the
page.

We now once again use `#issue-summary` as the view's element, with the
selectors inside adjusted accordingly.

Testing Done:
Unit tests pass.

Tested loading, interacting with, and reloading the table.

Reviewed at https://reviews.reviewboard.org/r/10426/