Return the correct review update user from the last update resource
3d1a32c9d7 changed the last update resource to return the correct author of an
update when the change was published by a different user. Unfortunately this
made the author for review changes be the review request submitter and not the
review submitter.

Testing Done:
Unit tests.

Also did a manual test where I created a review request with one user in window A, then posted a review with another in window B and returned to window A and verified that the updated bubble was correct (it wasn't before this change).

Reviewed at https://reviews.reviewboard.org/r/10422/