Add new color scheme definitions.
This change adds the new definitions for the Review Board color scheme.
This includes varying shades of grey, blue, brown, green, red, and
yellow. This doesn't completely cover all the colors used in the
product, but it's the vast majority of them. We have a bunch of
hard-coded colors throughout the CSS, many of which are slightly varying
shades of the same thing. These new definitions will allow us to keep
more consistency throughout.

Reviewed at https://reviews.reviewboard.org/r/10442/