Start converting main page colors over to new definitions.
This change starts to migrate the main page colors over to the new color
definitions. The bulk of this change relates to the `.page-colors`
mixin, which controls the color scheme of the header bar. It also
changes the `link-color` definition to use `blue-20`, which is a dark
teal-like color (though there are a number of links throughout which do
not yet use that style definition).

Testing Done:
Looked at the different page styles (base, reviewable, admin) and
verified that the topbar looked correct.

Reviewed at https://reviews.reviewboard.org/r/10442/