Add common styling for avatars.
We previously styled avatars differently in different parts of the UI.
Most places would style the avatars to be rounded, but some places would
not, and it entirely depended on where the avatar was.

This adds some common styling that gives avatars a rounded look by
default. Places that need to show a square view (such as the infobox)
can (and currently do) override the `border-radius` to remove the
rounded look.

Testing Done:
Looked around the UI for all places where we use avatars, both in
Review Board and RBCommons, checking that they appeared as expected.

Reviewed at https://reviews.reviewboard.org/r/10445/