Fix a regression with rendering using the fallback avatar.
The fallback avatar expected int sizes, so it could perform numeric
operations on them, but we weren't guaranteeing that we'd send ints
around. The `avatar` template tag would sometimes get string values
that it'd happily pass along.

We now convert the value to an int if it's a string, and gracefully fail
to render the avatar if that does not work.

Unit tests were added to test this functionality and to also fix
regressions due to changes in the avatar templates in Djblets.