Fix a unit test regression for WebHooks.
A while back, JSON serialization for the API was changed to always
return ordered keys. This affected WebHooks, but the unit test was never
fixed up to account for this.

This change fixes up the expected WebHook payload, ordering the keys.
When I say "ordering," it's based on the `json` module's method of
ordering, which orders them before converting to strings, which is why
`"null"` (`None`) is first.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10457/