Fix the generation of raw diffs for multi-commit review requests.
When grabbing the raw diff of a `DiffSet` on a multi-commit review
request, the diff would contain the cumulative diff and the diffs of all
files across all commits. This was due to the code attempting to gather
all `FileDiff`s pointing to the `DiffSet`, which included those from
commits.

This change updates the logic to be smarter, inspecting what we're
working with and fetching the files based on the capabilities of the
parent object being passed in.

Testing Done:
Unit tests passed.

Manually testing that I was only getting the cumulative diff when using
the diffs API or when using the "Download Diff" link.

Reviewed at https://reviews.reviewboard.org/r/10456/