Make draft colors consistent.
We had a bunch of different definitions of the "draft" color, several of
which were actually closer to the "ship-it" color than the draft banner
color. This change makes it all consistent throughout, creating a single
`@draft-color` definition which is based on the new color scheme. This
is then used everywhere that we want to indicate that there's something
in a draft state.

Some of this looked weird at first, because brains hate change. As I've
spent a little more time with it I've quickly gotten used to it and I
like it better.

Testing Done:
Looked throughout the UI at all the different draft-related items.

Reviewed at https://reviews.reviewboard.org/r/10466/