Make link colors consistent.
We had a bunch of different places where we had link colors of various
shades of blue. This standardizes them all into a single `@link-color`
definition, which is the same as `@blue-20`. While doing this, I decided
that the definition for that particular color (drawn from a mockup we
have) wasn't really blue enough to provide real contrast against white
and grey backgrounds, so I pushed the saturation a little bit.

Testing Done:
Flipped through the UI and looked for all the links I could find.

Reviewed at https://reviews.reviewboard.org/r/10467/