Update management commands for Django 1.11 compatibility and modern standards.
This updates all of the management commands to be compatible with both
Django 1.6 and 1.11, using the compatibility layer in Djblets. The major
differences between these releases is that there's only a `BaseCommand`
class, rather than a `NoArgsCommand` subclass (which wasn't even
important in Django 1.6), and that option parsing uses `argparse`
instead of `optparse` (which the Djblets compatibility class bridges for
us).

Many of the commands have also been updated to meet modern coding
standards. This includes doc strings, localization, and small style
tweaks. Not all of the commands received this treatment. Some are
scheduled to be removed, and others (like `fill-database`) requires a
larger amount of work than was intended for this change.

Testing Done:
Tested each management command individually on Django 1.6 and 1.11.

Reviewed at https://reviews.reviewboard.org/r/10476/