Fix a unit test regression from the fallback avatar changes in Djblets.
This is a trivial change that updates the expected response for the
fallback avatar service. Prior to the service, this test would expect
the results to be simply empty (since no service would be there to
provide a result). With the introduction of the fallback service, the
result is now a list of empty URLs.