Fix regressions from the json_dumps sorting change.
The recent change to `json_dumps` in Djblets to sort keys broke a couple
of unit tests in Review Board, which hard-coded results using the old
ordering. This is a simple change that updates the tests to compare
against the sorted payloads.

Testing Done:
Unit tests pass on Django 1.6 and 1.11 (with other in-progress changes).

Reviewed at https://reviews.reviewboard.org/r/10484/