Update the related object widgets for some Django 1.11 changes.
The related object widgets call `build_attrs` in order to create some
attributes to pass into the HTML for the widgets. Django's changed the
signature for this method since Django 1.6, which led to some crashes
during rendering.

The function itself is really simple in its purpose, so rather than
working around the changes, we just build the attributes dictionary in
the widget directly, fixing that compatibility problem.

Testing Done:
Unit tests pass on Django 1.6 and 1.11 (with other in-progress changes).

Reviewed at https://reviews.reviewboard.org/r/10486/