Update remaining URLs to use proper resolution and views.
Most URLs in the product properly resolved using registered URL names,
and registered by specifying the view itself rather than a string
representing the path to the view. We had a few that haven't yet been
updated, though. Namely, some URL building/references to administrative
URLs, and the static media views.

This updates the admin URL references to use URL names, and adds a
missing URL name for the main admin dashboard. It also updates the URL
registrations for the debug-only `/static/` and `/media/` URLs to
specify the view itself instead of a string representing the view.

Testing Done:
Unit tests pass on Django 1.6 and 1.11 (with other in-progress changes).

Tested the resulting URLs in the affected templates, and that both
static media and uploaded media work.

Reviewed at https://reviews.reviewboard.org/r/10489/