Fix unit tests on newer versions of Django Haystack.
Newer versions of Django Haystack changed the location of the variable
for retrieving the current signal processor. While the search code
itself is not sensitive to this, the unit tests are.

This change adds a function for retrieving the signal processor,
grabbing it from Haystack's `AppConfig` (if present) or using the old
location on the `haystack` module.

Testing Done:
Unit tests pass on Django 1.6 and 1.11 (with other in-progress changes).

Reviewed at https://reviews.reviewboard.org/r/10491/