Remove the Google Code hook instructions.
Google Code's gone, and we really need to hide the option here (we can't
outright remove it or some users with older servers might hit issues).
However, for now, we also need to not hit any errors when attempting to
render hook instructions on Django 1.11. This change removes the code
that renders this, rather than ports it over to the compatibility
methods.

Testing Done:
Verified that the button for hooks instructions is gone.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10514/